requirements.txt
# PyQT
python-mpd
